% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_explorer.R
\name{file_explorer}
\alias{file_explorer}
\title{file_explorer}
\usage{
file_explorer(
  viewer = getOption(x = "servers.tools.viewer", default = "dialog")
)
}
\arguments{
\item{viewer}{Where to display the gadget: \code{"dialog"},
\code{"pane"} or \code{"browser"} (see \code{\link[shiny]{viewer}}).}
}
\value{
UI gadget
}
\description{
file_explorer
}
\examples{
\dontrun{
library(shiny)
library(shinyFiles)
# shinyFiles::shinyFilesExample()
file_explorer()
}
}
