% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_explorer.R
\name{listFiles}
\alias{listFiles}
\title{List file in subdirectory}
\usage{
listFiles(path, maxDepth = 4, currentDepth = 1)
}
\arguments{
\item{path}{path to explore}

\item{maxDepth}{max depth to explore folders}

\item{currentDepth}{opened depth}
}
\value{
Tree list object
}
\description{
List file in subdirectory
}
