% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servers.tools-package.R
\docType{package}
\name{servers.tools-package}
\alias{servers.tools}
\alias{servers.tools-package}
\title{servers.tools: Extra Tools For Server Use}
\description{
Use tools.
}
\author{
\strong{Maintainer}: Sebastien Rochette \email{sebastien@thinkr.fr} (\href{https://orcid.org/0000-0002-1565-9313}{ORCID})

Other contributors:
\itemize{
  \item ThinkR [copyright holder]
}

}
\keyword{internal}
